backend/requirements.txt
fastapi==0.115.6
langchain==0.3.14
langchain_chroma==0.1.4
langchain_core==0.3.29
langchain_openai==0.2.14
langgraph==0.2.61
openai==1.59.3
passlib==1.7.4
psycopg2_binary==2.9.10
pydantic==2.10.4
PyJWT==2.10.1
python-dotenv==1.0.1
typing_extensions==4.12.2
uvicorn==0.34.0
